Move thread-related enums to reside with thread model
